# specs/001-digital-library-forum/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Library Forum Backend API
  version: 0.1.0
  description: REST contract for the digital library, academic forum, study groups, moderation, administration, statistics, and system configuration backend.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201': { description: Account created, content: { application/json: { schema: { $ref: '#/components/schemas/UserResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in and receive access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200': { description: Tokens issued, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '423': { description: Account locked }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200': { description: Tokens refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokenResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token
      responses:
        '204': { description: Logged out }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /documents/{documentId}:
    get:
      tags: [LibraryDocuments]
      security: []
      summary: View public or accessible document details
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document details, content: { application/json: { schema: { $ref: '#/components/schemas/DocumentResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /lecturer/documents:
    get:
      tags: [LecturerDocuments]
      summary: List lecturer-owned documents and review status
      responses:
        '200': { description: Lecturer documents }
    post:
      tags: [LecturerDocuments]
      summary: Upload a lecturer document for review
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201': { description: Document created pending review, content: { application/json: { schema: { $ref: '#/components/schemas/DocumentResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
  /lecturer/documents/{documentId}:
    patch:
      tags: [LecturerDocuments]
      summary: Update own lecturer document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document updated }
    delete:
      tags: [LecturerDocuments]
      summary: Delete own lecturer document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204': { description: Document deleted }
  /lecturer/documents/{documentId}/hide:
    post:
      tags: [LecturerDocuments]
      summary: Hide own document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document hidden }
  /forum/posts:
    get:
      tags: [Forum]
      security: []
      summary: List forum posts
      responses:
        '200': { description: Forum posts }
    post:
      tags: [Forum]
      summary: Create forum post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateForumPostRequest'
      responses:
        '201': { description: Post created }
  /forum/posts/{postId}:
    get:
      tags: [Forum]
      security: []
      summary: View forum post details
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200': { description: Post details }
    patch:
      tags: [Forum]
      summary: Edit own forum post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200': { description: Post updated }
    delete:
      tags: [Forum]
      summary: Delete own forum post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204': { description: Post deleted }
  /forum/posts/{postId}/comments:
    post:
      tags: [Forum]
      summary: Comment on forum post
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201': { description: Comment created }
  /forum/reports:
    post:
      tags: [Forum]
      summary: Report forum content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateForumReportRequest'
      responses:
        '201': { description: Report created }
...
  /content/documents/{documentId}/approve:
    post:
      tags: [ContentManagement]
      summary: Approve pending document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200': { description: Document approved }
  /content/documents/{documentId}/reject:
    post:
      tags: [ContentManagement]
      summary: Reject pending document with reason
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectDocumentRequest'
      responses:
        '200': { description: Document rejected }
        '400': { $ref: '#/components/responses/ValidationError' }
  /content/forum/posts/{postId}:
    delete:
      tags: [ContentManagement]
      summary: Delete violating forum post
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '204': { description: Post deleted }
  /content/forum/posts/{postId}/lock:
    post:
      tags: [ContentManagement]
      summary: Lock forum topic
      parameters:
        - $ref: '#/components/parameters/PostId'
      responses:
        '200': { description: Topic locked }
...
  /content/reports/{reportId}/handle:
    post:
      tags: [ContentManagement]
      summary: Handle violation report
      parameters:
        - $ref: '#/components/parameters/ReportId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandleReportRequest'
      responses:
        '200': { description: Report handled }
  /admin/users:
    get:
      tags: [AdminManagement]
      summary: Search users
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200': { description: Users }
  /admin/users/{userId}:
    get:
      tags: [AdminManagement]
      summary: View account details
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: User details }
    patch:
      tags: [AdminManagement]
      summary: Update account status
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: User updated }
  /admin/users/{userId}/lock:
    post:
      tags: [AdminManagement]
      summary: Lock account
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Account locked }
  /admin/users/{userId}/unlock:
    post:
      tags: [AdminManagement]
      summary: Unlock account
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Account unlocked }
  /admin/users/{userId}/roles:
    put:
      tags: [AdminManagement]
      summary: Replace user roles
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200': { description: Roles updated }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId: { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
    PostId: { name: postId, in: path, required: true, schema: { type: string, format: uuid } }
...
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    ReportId: { name: reportId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, fullName, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        fullName: { type: string, minLength: 1, maxLength: 120 }
        role: { type: string, enum: [STUDENT, LECTURER] }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    AuthTokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: '#/components/schemas/UserResponse' }
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        fullName: { type: string }
        status: { type: string }
        roles: { type: array, items: { type: string } }
...
    DocumentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { type: string }
        ownerId: { type: string, format: uuid }
...
    CreateDocumentRequest:
      type: object
      required: [title, file]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string }
        categoryId: { type: string, format: uuid }
        metadata: { type: object }
        file: { type: string, format: binary }
...
    CreateForumPostRequest:
      type: object
      required: [title, content]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        content: { type: string, minLength: 1 }
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
    CreateForumReportRequest:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType: { type: string, enum: [POST, COMMENT] }
        targetId: { type: string, format: uuid }
        reason: { type: string, minLength: 1 }
        description: { type: string }
...
    RejectDocumentRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 5, maxLength: 1000 }
    HandleReportRequest:
      type: object
      required: [status, resolutionNote]
      properties:
        status: { type: string, enum: [RESOLVED, REJECTED] }
        resolutionNote: { type: string, minLength: 1, maxLength: 2000 }
    AssignRolesRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        errors:
          type: array
          items: { type: string }